% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_gender.R
\name{bucket_gender}
\alias{bucket_gender}
\title{Create recoded gender column}
\usage{
bucket_gender(x, var, new_var = gender_bucket)
}
\arguments{
\item{x}{Data.frame containing gender data}

\item{var}{Original variable containing gender data. Must be specified}

\item{new_var}{Name of new gender variable. Defaults to gender_bucket}
}
\value{
A data.frame
}
\description{
This function takes a column of gender values and, in a new column,
returns the categorized values included in gender_walk.
}
\examples{
head(test_data) \%>\% bucket_gender(gender) \%>\% select(gender, gender_bucket)

}
