% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gender.R
\name{clean_gender}
\alias{clean_gender}
\title{Create recoded gender column}
\usage{
clean_gender(x, var, new_var = gender_clean)
}
\arguments{
\item{x}{Data.frame containing gender data}

\item{var}{Original variable containing gender data. Must be specified}

\item{new_var}{Name of new gender variable. Defaults to gender_clean}
}
\value{
A data.frame
}
\description{
Deprecated - please use bucket_gender instead. This function takes a column of gender values and, in a new column,
returns the categorized values included in gender_walk.
}
\examples{
head(test_data) \%>\% clean_gender(gender) \%>\% select(gender, gender_clean)

}
