% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_table_format.R
\name{demo_table_format}
\alias{demo_table_format}
\title{Format demographic report tables}
\usage{
demo_table_format(
  data,
  char_var = Characteristic,
  chars = c("Gender", "Race", "Region", "Companion Animals", "Went Fishing",
    "Handled Chickens"),
  char_header_blank = FALSE,
  stripe_fill = "lightblue",
  stripe_color = "white",
  text_color = "darkgray",
  return_html = FALSE,
  include_css = TRUE,
  ...
)
}
\arguments{
\item{data}{Required. A dataframe to be formatted as an html table. This dataframe should have the desired column names, with the exception of the main characteristic variable name, which can be overriden to blank.}

\item{char_var}{Required. Variable that contains characteristics. Set to Characteristic by default.}

\item{chars}{A vector of quoted characteristic groups. Table will be ordered in the same order these groups are listed.
These should NOT include specific traits (e.g., "Woman"), but instead, only the groups those traits fall into (e.g. "Gender").
These are the rows that will be fill in a different color.}

\item{char_header_blank}{If TRUE, the column specified in char_var will be set to blank. FALSE by default.}

\item{stripe_fill}{Color to fill characteristic row. "lightblue" by default.}

\item{stripe_color}{Color of text in characteristic row. "white" by default.}

\item{text_color}{Color of text in non-characteristic rows and header. "darkgray" by default.}

\item{return_html}{If TRUE, returns raw HTML of table. FALSE by default}

\item{include_css}{If TRUE, returns inline CSS for table formatting. TRUE by default. This is only returned if return_html is also TRUE}

\item{...}{Other arguments}
}
\value{
An HTML table or raw HTML
}
\description{
Characteristics can be reordered using the chars argument, but traits within those groups will be ordered as they are in the data.
The exception to this is for repeated options (e.g., "Yes" and "No"), which will be ordered in the way they first appear in the data.
}
\examples{
demo_table_format(df,
char_var = Trait,
chars = c("Income", "Education", "State", "Gender"))
}
