% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_maker.R
\name{donut_maker}
\alias{donut_maker}
\title{Make donut graph}
\usage{
donut_maker(
  x,
  color = "green",
  empty_color = "lightgray",
  label = NULL,
  label_color = NULL
)
}
\arguments{
\item{x}{Percentage to make donut graph of. Must be a numeric in decimal form.}

\item{color}{Fill color of graph. Green by default.}

\item{empty_color}{Color for remaining area. Light gray by default.}

\item{label}{Text for middle of graph. By default, x written as a percentage. Set to "" for blank.}

\item{label_color}{Color of text for middle of graph. By default, this is the same as color.}
}
\value{
A ggplot graph
}
\description{
Make donut graph
}
\examples{
donut_maker(.25, "red")
}
