% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fauna_blues.R
\name{fauna_blues}
\alias{fauna_blues}
\title{Return shades of blue based on the Faunalytics color palette}
\usage{
fauna_blues(x)
}
\arguments{
\item{x}{An integer greater than 0.}
}
\value{
A vector of hex codes
}
\description{
This function takes an integer input and returns a corresponding number of
hex codes based on the shades of blue in the Faunalytics color palette. In
other words, you give it a number of shades of blue you need it, it gives
you those shades.
When given 1, it returns Faunalytics blue. When given 2, it returns 
Faunalytics dark blue and light blue. When given 3, it returns Faunalytics 
dark blue, blue, and light blue. All values greater than 3 are interpolated 
between Faunalytics dark blue and light blue, which will be the darkest and 
lightest colors returned.
}
\examples{
barplot(seq(1,5,1), col = fauna_blues(5))
}
