% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fauna_theme.R
\name{fauna_theme}
\alias{fauna_theme}
\title{Apply Faunalytics aesthetics to ggplot graphs}
\usage{
fauna_theme()
}
\arguments{
\item{p}{A ggplot object}
}
\value{
A Faunalytics-themed ggplot object
}
\description{
Apply Faunalytics aesthetics to ggplot graphs
}
\examples{
ggplot(mtcars, aes(x = cyl, y = mpg)) +
geom_col() +
fauna_theme()
}
