% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fauna_theme_flip.R
\name{fauna_theme_flip}
\alias{fauna_theme_flip}
\title{Apply Faunalytics aesthetics to ggplot graphs that use coord_flip}
\usage{
fauna_theme_flip()
}
\value{
A Faunalytics-themed ggplot object
}
\description{
Apply Faunalytics aesthetics to ggplot graphs that use coord_flip
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = cyl, y = mpg)) +
    geom_col() +
    fauna_theme_flip()
}
