% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_walk.R
\name{gender_walk}
\alias{gender_walk}
\title{Gender recoding data}
\format{
A data frame with 20 rows and 2 variables:
\describe{
  \item{original}{open-ended response}
  \item{clean}{Faunalytics category}
}
}
\description{
A dataset containing open-ended responses to the question "What is your gender?"
and the category each falls into for consistency.
}
