% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_full_palette.R
\name{return_full_palette}
\alias{return_full_palette}
\title{Return all colors recognized by fauna_colors function}
\usage{
return_full_palette()
}
\value{
Named vector of all colors recognized by fauna_colors
}
\description{
Return all colors recognized by fauna_colors function
}
\examples{
return_full_palette()
}
