% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_html.R
\name{return_html}
\alias{return_html}
\title{Return HTML of table}
\usage{
return_html(x, include_css = TRUE, write = FALSE, path = "table.txt")
}
\arguments{
\item{x}{A table object generated by table_format. Note that the return_html argument of table_format must be set to FALSE (default) for this function to work.}

\item{include_css}{If TRUE, returns inline CSS for table formatting. TRUE by default. This is only returned if return_html is also TRUE}

\item{write}{If TRUE, write results to the file specified in the path argument. FALSE by default.}

\item{path}{File path to be written to if write is TRUE. "table.txt" in working directory by default.}
}
\value{
The HTML of the table provided.
}
\description{
Return HTML of table
}
\examples{
table_format(cars) \%>\% return_html(write = TRUE, path = "output/Figure1.txt")
}
