% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_val.R
\name{square_val}
\alias{square_val}
\title{Square a value}
\usage{
square_val(x)
}
\arguments{
\item{x}{A numeric input to be squared}
}
\value{
A number
}
\description{
This function takes a numeric value and squares it.
It is intended to be used as a replacement for 'value^2'
}
\examples{
square_val(4)
}
