% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_to_region.R
\name{state_to_region}
\alias{state_to_region}
\title{Create region variable from state input}
\usage{
state_to_region(
  data,
  state.var = state,
  region.var = region,
  factorize = TRUE,
  region.levels = c("Northeast", "Midwest", "South", "West")
)
}
\arguments{
\item{data}{A data frame or data frame extension (e.g. a tibble)}

\item{state.var}{Variable containing state values. Defaults to state}

\item{region.var}{Variable to be created containing regions. Defaults to region}

\item{factorize}{A logical evaluation to TRUE or FALSE indicating whether to make the region.var a factor. Defaults to TRUE}

\item{region.levels}{An optional vector of the unique values (as character strings) that region.var might have taken. Defaults to c("Northeast", "Midwest", "South", "West")}
}
\value{
A data frame
}
\description{
This function works with title case names of the 50 states and D.C., as well
as with all caps 2-letter abbreviations with and without periods (e.g., NJ and N.J.).
Default regions are defined by the U.S. Census.
}
\examples{
USArrests |> tibble::rownames_to_column("us_state") |>
state_to_region(us_state, us_region)
}
