% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tibble.R
\name{summary_tibble}
\alias{summary_tibble}
\title{Make a summary tibble out of model or model summary object}
\usage{
summary_tibble(x, stars = TRUE, clean = FALSE, p_name = NULL)
}
\arguments{
\item{x}{Model or model summary object.}

\item{stars}{Create 'stars' column containing asterisks indicating significance.}

\item{clean}{Clean names using janitor::clean_names. FALSE by default.}

\item{p_name}{Optional. Name of p-value column. If p-values are not in a 
column named either "Pr(>|t|)" or "Pr(>|z|)", the `stars` argument will not 
work without this argument.}
}
\value{
A tibble
}
\description{
This function takes a lm or glm model or model summary object and returns a 
tibble with key measures of interest (e.g., coefficient estimate, p-values).
These measures may change depending on the type of model.
}
\examples{
m1 <- lm(mpg ~ cyl + hp + wt, data = mtcars)
summary_tibble(m1)
}
