% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_format.R
\name{table_format}
\alias{table_format}
\title{Format report tables}
\usage{
table_format(
  data,
  header_fill = "blue",
  header_color = "white",
  cell_fill = "white",
  text_color = "darkgray",
  border_color = "white",
  shade_color = "lightgray",
  na.rm = TRUE,
  h_aligns = NULL,
  return_html = FALSE,
  include_css = TRUE,
  write = FALSE,
  path = "table.txt",
  ...
)
}
\arguments{
\item{data}{Required. A dataframe to be formatted as an html table}

\item{header_fill}{Color of header background. Blue by default}

\item{header_color}{Color of header text. White by default}

\item{cell_fill}{Color of table body cells. White by default}

\item{text_color}{Color of table body text. Dark gray by default}

\item{border_color}{Color of cell borders. White by default}

\item{shade_color}{Color of alternate row shading. Gray by default.
Currently, the only other option is "lightblue", which will make text in those rows white.}

\item{na.rm}{Remove NA values from character columns and replace with blanks. TRUE by default.
If FALSE, NA will show up in any cells where it appears in the data you feed into this function.}

\item{h_aligns}{Horizontal alignment of columns. If this is not specified, R will guess.
You can either specify a single string which will be applied to all columns,
or a vector of strings where that vector's length is equal to the number of columns in the data.
Options must be one of: "left", "center", "right"}

\item{return_html}{If TRUE, returns raw HTML of table. FALSE by default}

\item{include_css}{If TRUE, returns inline CSS for table formatting. TRUE by default. This is only returned if return_html is also TRUE}

\item{write}{If TRUE, write results to the file specified in the path argument. FALSE by default.}

\item{path}{File path to be written to if write is TRUE. "table.txt" in working directory by default.}

\item{...}{Other arguments}
}
\value{
An HTML table or raw HTML
}
\description{
Format a table using Faunalytics standards.
If you want to specify column widths, leave return_html as FALSE and pipe the results of table_format
into cols_width, where you will specify the column widths. To return html, pipe the result of that pipeline
into return_html. For more, see ?cols_width and ?return_html. An example is given below.
}
\examples{
table_format(head(mtcars))
table_format(head(cars)) \%>\% return_html()
}
