% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_adj.R
\name{bh_adj}
\alias{bh_adj}
\title{Apply Benjamini-Hochberg adjustment to model summary object}
\usage{
bh_adj(x, fdr = 0.05, p_name = "Pr(>|t|)", coef_vec = NULL)
}
\arguments{
\item{x}{Model summary object. For example, summary(model).}

\item{fdr}{False discovery rate. 0.05 by default.}

\item{p_name}{Name of p-value column in the summary object. "Pr(>|t|)" by
default.}

\item{coef_vec}{Optional. A vector of coefficient names as strings.
p-values associated with the coefficients in this vector will be included in 
the adjustment and all other coefficients will be excluded.}
}
\value{
A tibble
}
\description{
This function takes a model summary object (e.g., summary(model)) and
returns a tibble with the Benjamini-Hochberg adjustment for false discovery
rate applied. New columns include p-value rank (`rank`), adjusted alpha
(`adj_a`), and a logical indicator of significance with the 
Benjamini-Hochberg adjustment applied (`bh_sig`).
}
\examples{
m1 <- lm(mpg ~ cyl + hp + wt, data = mtcars)
sum_m1 <- summary(m1)
bh_adj(sum_m1)

}
