% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_table_format.R
\name{demo_table_format}
\alias{demo_table_format}
\title{Format demographic report tables}
\usage{
demo_table_format(
  data,
  char_var = Characteristic,
  chars = c("Gender", "Race", "Region", "Companion Animals", "Went Fishing",
    "Handled Chickens"),
  char_header_blank = FALSE,
  stripe_fill = "lightblue",
  stripe_color = "white",
  text_color = "darkgray",
  na.rm = T,
  return_html = FALSE,
  include_css = TRUE,
  write = FALSE,
  path = "table.txt",
  stripe_colour = NULL,
  text_colour = NULL,
  ...
)
}
\arguments{
\item{data}{Required. A dataframe to be formatted as an html table. This dataframe should have the desired column names, with the exception of the main characteristic variable name, which can be overriden to blank.}

\item{char_var}{Required. Variable that contains characteristics. Set to Characteristic by default.}

\item{chars}{A vector of quoted characteristic groups. Table will be ordered in the same order these groups are listed.
These should NOT include specific traits (e.g., "Woman"), but instead, only the groups those traits fall into (e.g. "Gender").
These are the rows that will be fill in a different color.}

\item{char_header_blank}{If TRUE, the column specified in char_var will be set to blank. FALSE by default.}

\item{stripe_fill}{Color to fill characteristic row. "lightblue" by default.}

\item{stripe_color}{Color of text in characteristic row. "white" by default.}

\item{text_color}{Color of text in non-characteristic rows and header. "darkgray" by default.}

\item{na.rm}{Remove NA values from character columns and replace with blanks. TRUE by default.
If FALSE, NA will show up in any cells where it appears in the data you feed into this function.}

\item{return_html}{If TRUE, returns raw HTML of table. FALSE by default}

\item{include_css}{If TRUE, returns inline CSS for table formatting. TRUE by default. This is only returned if return_html is also TRUE}

\item{write}{If TRUE, write results to the file specified in the path argument. FALSE by default.}

\item{path}{File path to be written to if write is TRUE. "table.txt" in working directory by default.}

\item{stripe_colour}{See stripe_color}

\item{text_colour}{See text_color}

\item{...}{Other arguments}
}
\value{
An HTML table or raw HTML
}
\description{
Characteristics can be reordered using the chars argument, but traits within those groups will be ordered as they are in the data.
The exception to this is for repeated options (e.g., "Yes" and "No"), which will be ordered in the way they first appear in the data.
}
\examples{
temp_data <- data.frame(
"Trait" = c("Gender", "Woman or Other", "Man",
    "Income", "Less than $25,000", "$25,000 to $49,999", "$50,000 to $74,999",
    "$75,000 to $99,999", "$100,000 or more",
    "Education", "Less than high school graduate",
    "High school graduate", "Some college or university, no degree",
    "Associate's degree or college diploma", "Bachelor's degree",
    "Master's degree", "Professional or doctoral degree"),
"Percent" = c("", "49\%", "51\%",
    "", "12\%", "20\%", "19\%",
    "22\%", "27\%",
    "", "3\%",
    "18\%", "17\%",
    "18\%", "32\%",
   "10\%", "3\%"))
demo_table_format(temp_data,
char_var = Trait,
chars = c("Gender", "Income", "Education"))
}
