% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_format.R
\name{table_format}
\alias{table_format}
\title{Format report tables}
\usage{
table_format(
  data,
  header_fill = "blue",
  header_color = "white",
  cell_fill = "white",
  text_color = "darkgray",
  border_color = "white",
  shade_color = "lightgray",
  return_html = FALSE,
  include_css = TRUE,
  ...
)
}
\arguments{
\item{data}{A dataframe to be formatted as an html table}

\item{header_fill}{Color of header background. Blue by default}

\item{header_color}{Color of header text. White by default}

\item{cell_fill}{Color of table body cells. White by default}

\item{text_color}{Color of table body text. Dark gray by default}

\item{border_color}{Color of cell borders. White by default}

\item{shade_color}{Color of alternate row shading. Gray by default.
Currently, the only other option is "lightblue", which will make text in those rows white.}

\item{return_html}{If TRUE, returns raw HTML of table. FALSE by default}

\item{include_css}{If TRUE, returns inline CSS for table formatting. TRUE by default. This is only returned if return_html is also TRUE}

\item{...}{Other arguments}
}
\value{
An HTML table or raw HTML
}
\description{
Format report tables
}
\examples{
table_format(head(mtcars))
}
